增加操作日志查询接口

+++ web/src/main/java/net/wuxianjie/web/service/impl/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package net.wuxianjie.web.service.impl;
 
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import net.wuxianjie.core.exception.BadRequestException;
 import net.wuxianjie.core.exception.DataConflictException;
@@ -9,14 +10,16 @@ import net.wuxianjie.web.controller.UserController;
 import net.wuxianjie.web.mapper.UserMapper;
 import net.wuxianjie.web.model.Account;
 import net.wuxianjie.web.model.User;
-import net.wuxianjie.web.model.WroteDb;
+import net.wuxianjie.web.model.Wrote2Database;
+import net.wuxianjie.web.service.OperationLogService;
 import net.wuxianjie.web.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -28,11 +31,12 @@ import java.util.List;
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class UserServiceImpl implements UserService {
 
+  private final OperationLogService logService;
   private final UserMapper userMapper;
   private final PasswordEncoder passwordEncoder;
 
   @Override
-  public PaginationData<List<User>> loadUsers(@NotNull final Integer pageNo, @NotNull final Integer pageSize, final String fuzzyUsername) {
+  public PaginationData<List<User>> loadUsers(@NonNull final Integer pageNo, @NonNull final Integer pageSize, final String fuzzyUsername) {
     final List<User> users = userMapper.findUsersPaginationByUsername(pageNo * pageSize, pageSize, fuzzyUsername);
     final int total = userMapper.findUserCountByUsername(fuzzyUsername);
     return new PaginationData<>(total, pageNo, pageSize, users);
@@ -40,7 +44,7 @@ public class UserServiceImpl implements UserService {
 
   @Override
   @Transactional(rollbackFor = Exception.class)
-  public WroteDb saveUser(@NotNull final UserController.UserToAdd userToAdd) {
+  public Wrote2Database saveUser(@NonNull final UserController.UserToAdd userToAdd) {
     // 若已存在同名用户，则拒绝添加
     final int count = userMapper.findUserCountByUsername(userToAdd.getUsername());
 
@@ -55,12 +59,16 @@ public class UserServiceImpl implements UserService {
     // 入库
     final int addedNum = userMapper.addUser(userToAdd);
 
-    return new WroteDb(addedNum, "新增用户成功");
+    // 记录日志
+    final String logMessage = String.format("新增用户名为%s的用户", userToAdd.getUsername());
+    logService.saveOperationLog(LocalDateTime.now(), logMessage);
+
+    return new Wrote2Database(addedNum, "新增用户成功");
   }
 
   @Override
-  @Transactional
-  public WroteDb updateUser(@NotNull final UserController.UserToUpdate userToUpdate) {
+  @Transactional(rollbackFor = Exception.class)
+  public Wrote2Database updateUser(@NonNull final UserController.UserToUpdate userToUpdate) {
     // 若库中不存在该用户ID的数据，则直接退出方法
     final Account account = userMapper.findAccountByUserId(userToUpdate.getUserId());
 
@@ -68,20 +76,28 @@ public class UserServiceImpl implements UserService {
       throw new BadRequestException("用户ID不存在");
     }
 
-    // 判断数据是否需要更新，并将不需要更新的数据设置为null
-    final boolean shouldUpdate = shouldUpdateAndSetNull4NotNeeded(account, userToUpdate);
+    // 判断数据是否需要更新
+    final boolean shouldUpdate = shouldUpdateUserInfo(account, userToUpdate);
+    // 将不需要更新的数据设置为null并返回详细日志
+    final String updatedLog = makeNull4NotNeedUpdatedFiledAndReturnLog(account, userToUpdate);
 
-    // 入库
     if (shouldUpdate) {
+      // 入库
       final int updatedNum = userMapper.updateUser(userToUpdate);
-      return new WroteDb(updatedNum, "更新用户成功");
+
+      // 记录日志
+      final String logMessage = String.format("更新用户信息：%s", updatedLog);
+      logService.saveOperationLog(LocalDateTime.now(), logMessage);
+
+      return new Wrote2Database(updatedNum, "更新用户成功");
     }
 
-    return new WroteDb(0, "无需更新用户");
+    return new Wrote2Database(0, "无需更新用户");
   }
 
   @Override
-  public WroteDb updatePassword(@NotNull final UserController.PasswordToUpdate passwordToUpdate) {
+  @Transactional(rollbackFor = Exception.class)
+  public Wrote2Database updatePassword(@NonNull final UserController.PasswordToUpdate passwordToUpdate) {
     // 若库中不存在该用户ID的数据，则直接退出方法
     final Account account = userMapper.findAccountByUserId(passwordToUpdate.getUserId());
 
@@ -90,7 +106,7 @@ public class UserServiceImpl implements UserService {
     }
 
     // 若旧密码与库中的密码不匹配，则直接退出方法
-    final boolean isRightOldPassword = passwordEncoder.matches(passwordToUpdate.getOldPassword(), account.getAccountPassword());
+    final boolean isRightOldPassword = passwordEncoder.matches(passwordToUpdate.getOldPassword(), account.getPassword());
 
     if (!isRightOldPassword) {
       throw new BadRequestException("旧密码错误");
@@ -102,40 +118,71 @@ public class UserServiceImpl implements UserService {
     // 入库
     final int updatedNum = userMapper.updatePassword(passwordToUpdate.getUserId(), encodedNewPassword);
 
-    return new WroteDb(updatedNum, "修改密码成功");
+    // 记录日志
+    final String logMessage = String.format("修改用户名为%s的密码", account.getName());
+    logService.saveOperationLog(LocalDateTime.now(), logMessage);
+
+    return new Wrote2Database(updatedNum, "修改密码成功");
   }
 
   @Override
-  @Transactional
-  public WroteDb removeUser(final int userId) {
+  @Transactional(rollbackFor = Exception.class)
+  public Wrote2Database removeUser(final int userId) {
+    // 若用户ID不存在，则直接返回
+    final Account account = userMapper.findAccountByUserId(userId);
+
+    if (account == null) {
+      throw new BadRequestException("用户ID不存在");
+    }
+
+    // 入库
     final int deletedNum = userMapper.deleteUserByUserId(userId);
-    return new WroteDb(deletedNum, deletedNum == 0 ? "无需删除用户" : "删除用户成功");
+
+    // 记录日志
+    final String logMessage = String.format("删除用户名为%s的用户", account.getName());
+    logService.saveOperationLog(LocalDateTime.now(), logMessage);
+
+    return new Wrote2Database(deletedNum, "删除用户成功");
   }
 
-  /**
-   * 判断是否需要更新数据库中数据，并将不需要更新的字段设置为null
-   */
-  private boolean shouldUpdateAndSetNull4NotNeeded(final Account account, final UserController.UserToUpdate userToUpdate) {
-    boolean shouldUpdate = false;
+  private boolean shouldUpdateUserInfo(final Account account, final UserController.UserToUpdate userToUpdate) {
+    final boolean isSamePassword = userToUpdate.getPassword() != null
+        && passwordEncoder.matches(userToUpdate.getPassword(), account.getPassword());
 
-    final boolean isSamePassword = userToUpdate.getPassword() != null && passwordEncoder.matches(userToUpdate.getPassword(), account.getAccountPassword());
     if (userToUpdate.getPassword() != null && !isSamePassword) {
-      shouldUpdate = true;
+      return true;
+    }
 
+    final boolean isSameRoles = StringUtils.isNullEquals(account.getRoles(), userToUpdate.getRoles());
+
+    return userToUpdate.getRoles() != null && !isSameRoles;
+  }
+
+  private String makeNull4NotNeedUpdatedFiledAndReturnLog(final Account account, final UserController.UserToUpdate userToUpdate) {
+    final List<String> logBuilder = new ArrayList<>();
+
+    final boolean isSamePassword = userToUpdate.getPassword() != null
+        && passwordEncoder.matches(userToUpdate.getPassword(), account.getPassword());
+
+    if (userToUpdate.getPassword() != null && !isSamePassword) {
       // 编码明文密码
       final String encodedPassword = passwordEncoder.encode(userToUpdate.getPassword());
       userToUpdate.setPassword(encodedPassword);
-    } else {
+
+      logBuilder.add(String.format("重置了%s的密码", account.getName()));
+    } else if (userToUpdate.getPassword() != null){
       userToUpdate.setPassword(null);
     }
 
-    final boolean isSameRoles = StringUtils.isNullEquals(account.getAccountRoles(), userToUpdate.getRoles());
+    final boolean isSameRoles = StringUtils.isNullEquals(account.getRoles(), userToUpdate.getRoles());
+
     if (userToUpdate.getRoles() != null && !isSameRoles) {
-      shouldUpdate = true;
-    } else {
+      logBuilder.add(String.format("将角色从%s修改为%s",
+          account.getRoles(), userToUpdate.getRoles().isEmpty() ? "空" : userToUpdate.getRoles()));
+    } else if (userToUpdate.getRoles() != null) {
       userToUpdate.setRoles(null);
     }
 
-    return shouldUpdate;
+    return String.join("；", logBuilder);
   }
 }
